requirements.txt
httpx
python-dotenv
pydantic_ai
streamlit